Sales Data Analysis
This is not the first project I have started but rather the first one I completed. I still haven't done the visualization but that I will probably do it in PowerBI or maybe Tableau. It is a sales data file. I have learnt a new library (not completely), i.e. NLTK which is a natural language processing library in python